requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
aioconsole==0.7.1
aiohttp==3.9.5